pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minka"
version = "0.1.0"
description = "Asistente Virtual de Ciberseguridad - Proyecto UCM Master"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "UCM Cybersecurity Master Student"}
]
keywords = ["cybersecurity", "vulnerability-research", "ai-assistant", "github-copilot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "github-copilot-sdk>=0.1.20",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # CLI & UI
    "click>=8.1.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "questionary>=2.0.0",
    
    # HTTP & APIs
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    
    # Database & Cache
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "alembic>=1.12.0",
    
    # Security Tools
    "python-nmap>=0.7.1",
    "shodan>=1.31.0",
    "requests>=2.31.0",
    
    # Data Processing
    "pyyaml>=6.0.1",
    "jsonschema>=4.20.0",
    "python-dotenv>=1.0.0",
    
    # Logging & Monitoring
    "structlog>=23.0.0",
    "prometheus-client>=0.19.0",
    
    # Testing (included for convenience)
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "respx>=0.20.0",
    
    # Development
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "black>=23.0.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
]

voice = [
    "python-telegram-bot>=20.0",
    "speechrecognition>=3.10.0",
    "pyttsx3>=2.90",
]

dev = [
    "pre-commit>=3.5.0",
    "ipython>=8.17.0",
    "jupyter>=1.0.0",
]

[project.scripts]
minka = "src.cli.main:main"

[project.urls]
Homepage = "https://github.com/statick/minka"
Documentation = "https://minka.readthedocs.io"
Repository = "https://github.com/statick/minka"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
